autoexecjob bug fixed tail console

+++ src/main/java/codedriver/module/autoexec/api/job/action/AutoexecJobConsoleLogTailApi.java
@@ -46,7 +46,7 @@ public class AutoexecJobConsoleLogTailApi extends PrivateApiComponentBase {
 
     @Input({
             @Param(name = "jobId", type = ApiParamType.LONG, isRequired = true, desc = "作业Id"),
-            @Param(name = "runnerId", type = ApiParamType.INTEGER, isRequired = true, desc = "runnerId"),
+            @Param(name = "runnerId", type = ApiParamType.LONG, isRequired = true, desc = "runnerId"),
             @Param(name = "logPos", type = ApiParamType.LONG, isRequired = true, desc = "日志读取位置,-1:获取最新的数据"),
             @Param(name = "direction", type = ApiParamType.ENUM, rule = "up,down", isRequired = true, desc = "读取方向，up:向上读，down:向下读")
     })
